+++ crates/optimism/payload/src/lib.rs
@@ -16,5 +16,7 @@ pub use builder::OpPayloadBuilder;
 pub mod error;
 pub mod payload;
 pub use payload::{OpBuiltPayload, OpPayloadAttributes, OpPayloadBuilderAttributes};
+mod traits;
+pub use traits::*;
 
 pub mod config;
